Abstract Window handling to platform.h and handle rendering in Render.h
platform.h will handle any differences between Linux and Windows.
render.h will handle the common rendering code between the two,
utilizing the WindowConfig struct.

This can be done because the underlying code between Linux and Windows
simply rely on a framebuffer for pixel manipulation, so algorithms are
performing the same operations, and differences between them can be
handled elsewhere.

i.e its not that rendering is different between Windows and Linux, but
how you relay the Window changes to the OS, which should be handled in
platform and not in render