Refactor World, WorldHandler, and test, add engine_options.h
WorldHandler is back, and better than ever.
WorldHandlers are now children of WindowConfigs.
There is now a Vector of Worlds attached to WorldHandlers

There is now a more visible way that users of the engine can implement
their code, with an actual test folder that might actually work.

Everything is completely untested, and uncompiled :)