Interm shape commit development
Vertex struct is now floats, because i thought it was floats this entire
time

Currently a line can be drawn from a Square, because vertices aren't
sorted, or generated in order properly, since the 'graph' is inverted on
the Y Axis for the Framebuffer (0, 0) is top left.

world_handler now (attempts) to transpose the Shape's vertices to World
Poses based on game objects 'origin' (bottom left? of shape?), lots of
work to be fleshed out still.